Move blocks to separate files
